Organise files and refactor code and views
