+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\Naiky\\Documents\\GitHub\\StoreManager\\api\\endpoints\\__init__.py":[1,2,4],"C:\\Users\\Naiky\\Documents\\GitHub\\StoreManager\\api\\endpoints\\app.py":[1,2,3,5,7,9,14,19,24,28,33,38,62,81,86,96,123,128,140,161,174,175,202,207,98,100,101,102,103,105,108,109,111,113,116,119,120,106,16,26,30,164,166,169,170,171,167,11,131,133,134,114,117,142,144,145,148,150,154,155,157,158,151,125,137,35,177,179,180,183,185,188,190,192,196,198,199,204,193,210,212,215,21,213,186,40,42,43,44,45,47,50,52,55,58,59,56,83,88,89,93,90,53,64,66,67,69,70,72,74,75,48,78],"C:\\Users\\Naiky\\Documents\\GitHub\\StoreManager\\api\\endpoints\\functions.py":[1,2,3,5,7,8,9,12,14,20,39,40,45,62,64,70,80,85,102,15,16,17,18,22,23,24,25,28,30,31,36,41,42,43,48,49,50,51,52,54,55,58,59,34,65,66,67,68,71,72,74,78,81,88,89,90,91,94,96,97,82,83,75,103,104,105,108]}}
+!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\Naiky\\Documents\\GitHub\\StoreManager\\api\\views\\__init__.py":[1],"C:\\Users\\Naiky\\Documents\\GitHub\\StoreManager\\api\\views\\app.py":[1,2,3,4,6,8,10,12,14,19,24,29,33,38,21,31,35,16,40,26],"C:\\Users\\Naiky\\Documents\\GitHub\\StoreManager\\api\\views\\products_views.py":[1,2,4,6,34,39,51,72,8,10,11,12,13,15,19,20,22,24,27,30,31,16,75,77,80,81,82,78,42,44,45,25,28,53,55,56,59,61,65,66,68,69,62,36,48],"C:\\Users\\Naiky\\Documents\\GitHub\\StoreManager\\api\\views\\functions.py":[1,2,3,5,7,8,9,12,14,20,39,40,45,62,64,70,80,85,102,15,16,17,18,22,23,24,25,28,30,31,36,41,42,43,48,49,50,51,52,54,55,58,59,34,65,66,67,68,71,72,74,78,81,88,89,90,91,94,96,97,82,83,75,103,104,105,108],"C:\\Users\\Naiky\\Documents\\GitHub\\StoreManager\\api\\views\\sales_views.py":[1,2,4,6,7,34,39,9,11,12,15,17,20,22,24,28,30,31,36,25,42,44,47,45,18],"C:\\Users\\Naiky\\Documents\\GitHub\\StoreManager\\api\\views\\user_views.py":[1,2,4,6,30,49,54,8,10,11,12,13,15,18,20,23,26,27,24,51,56,57,61,58,21,32,34,35,37,38,40,42,43,16,46]}}
