+++ .travis.yml
@@ -7,7 +7,7 @@ install:
   - pip install -r requirements.txt
 
 script:
-  - py.test --cov api/endpoints/
+  - py.test --cov api/views
 
 after_success:
   - coveralls
