+++ api/tests/test_products.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from flask import json
-from api.endpoints.app import app
-from api.endpoints.functions import products
+from api.views.app import app
+from api.views.functions import products
 
 sample_product = {
     "product_name": "water",
