+++ requirements.txt
@@ -1,61 +1,18 @@
-aniso8601==3.0.2
-astroid==2.0.4
-atomicwrites==1.1.5
-attrs==18.1.0
-autoenv==1.0.0
-autopep8==1.3.5
-backcall==0.1.0
-certifi==2018.8.13
-chardet==3.0.4
-click==6.7
-colorama==0.4.0
 coverage==4.5.1
 coveralls==1.4.0
-decorator==4.3.0
-docopt==0.6.2
-flake8==3.5.0
 Flask==1.0.2
 Flask-HTTPAuth==3.2.4
 Flask-JWT-Extended==3.12.1
 Flask-RESTful==0.3.6
 Flask-WTF==0.14.2
 gunicorn==19.9.0
-idna==2.7
-ipython==6.5.0
-ipython-genutils==0.2.0
-isort==4.3.4
-itsdangerous==0.24
-jedi==0.12.1
 Jinja2==2.10
-lazy-object-proxy==1.3.1
-MarkupSafe==1.0
-mccabe==0.6.1
-more-itertools==4.3.0
-parso==0.3.1
-pickleshare==0.7.4
-pipenv==2018.7.1
-pluggy==0.7.1
-prompt-toolkit==1.0.15
 psycopg2==2.7.5
 py==1.5.4
-pycodestyle==2.4.0
-pyflakes==1.6.0
-Pygments==2.2.0
 PyJWT==1.6.4
-pylint==2.1.1
 pytest==3.7.1
 pytest-cov==2.6.0
-python-dotenv==0.9.1
-pytz==2018.5
 requests==2.19.1
-simplegeneric==0.8.1
-six==1.11.0
-traitlets==4.3.2
-urllib3==1.23
-virtualenv==16.0.0
-virtualenv-clone==0.3.0
-virtualenvwrapper-win==1.2.5
-wcwidth==0.1.7
 Werkzeug==0.14.1
 wrapt==1.10.11
 WTForms==2.2.1
